hdd_data_analysis_v2/requirements.txt
pandas>=1.5.0
sqlalchemy>=1.4.0
pyodbc>=4.0.0
schedule>=1.2.0
pyyaml>=6.0
psutil>=5.9.0
numpy>=1.24.0 